requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==21.2.1
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
click==8.0.4
constantly==15.1.0
cryptography==40.0.2
daphne==3.0.2
dataclasses==0.8
Deprecated==1.2.14
distlib==0.3.8
Django==3.2.25
django-ckeditor==6.3.2
django-cors-headers==3.10.1
django-js-asset==2.0.0
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
filelock==3.4.1
gunicorn==21.2.0
h11==0.13.0
h2==4.1.0
hiredis==2.0.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==22.10.0
jwcrypto==1.5.1
msgpack==1.0.5
numpy==1.19.5
oauthlib==3.2.2
openpyxl==3.1.2
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
platformdirs==2.4.0
priority==1.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
razorpay==1.4.2
redis==4.3.6
requests==2.27.1
service-identity==21.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.18
uvicorn==0.16.0
virtualenv==20.17.1
webencodings==0.5.1
wrapt==1.16.0
xlrd==2.0.1
zipp==3.6.0
zope.interface==5.5.2
